Changed program structure
I added separate files for the different classes. For every class there
is a header file (.hpp) and a “source” filee (.cpp). The idea is that
the function description and explanation is written in the header file
and that implementation is written in the source file. This makes (at
least I think so) the code a lot clearer and easier to read.